memset y otros

+++ ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: ide-ruit <ide-ruit@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 18:45:04 by ide-ruit          #+#    #+#             */
-/*   Updated: 2023/09/18 19:51:06 by ide-ruit         ###   ########.fr       */
+/*   Updated: 2023/09/18 20:13:55 by ide-ruit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
-	int	cont;
+	size_t	cont;
 	
 	cont = 0;
 	i = 0;
